chore: simplify make-dotenv.sh by not separately exporting vars
Prior to this change each variable defined in the if-local-else-if-fit
section were separately exported later in the file. This could make it
easy to forget to export variables, and was an unnecessary extra set of
lines of code. Instead this commit just exports the variables at the
time they're defined.